refactor(order): extract Features/ConfirmOrder + CancelOrder slices (#156)
